TripNavBar is moved outside of Dashboard so it can be reused in Expenses and Landmarks. Added a few mmore NavBar options
