+++ client/src/Components/expenseTracker/expenseEntry.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
 const ExpenseEntry = (props) => (
-  <div>{props.expense.description} - ${props.expense.amount}, paid for by {props.payer}</div>
+  <div class="ExpenseEntry">{props.expense.description} - ${props.expense.amount}, paid for by {props.payer}</div>
 )
 
 export default ExpenseEntry;
