+++ client/src/Components/expenseTracker/expenseTracker.jsx
@@ -1,9 +1,14 @@
 import React from 'react';
 import reducer from '../../Reducers';
+import { Row } from 'react-bootstrap';
+import { Col } from 'react-bootstrap';
+
 import ExpenseEntry from './expenseEntry.jsx';
 import ExpenseInput from './expenseInput.jsx';
 import { connect } from 'react-redux';
 import $ from 'jquery';
+import TripNavBar from '../tripDashboard/tripNavBar.jsx';
+import dummyData from '../tripDashboard/dummyData.js';
 
 const SERVER_URL = HOSTNAME;
 
@@ -72,24 +77,25 @@ class ExpenseTracker extends React.Component {
 
 	render() {
 		return(
-			<div className="expbody">
-        <h3>Expenses Tracker</h3>
-        <div>
-          <ExpenseInput usersOnTrip={this.state.usersOnTrip} fetchExpenses={this.fetchExpenses.bind(this)} />
-          <hr />
+			<Row> 
+        <Col md={8} mdOffset={2}>
+          <TripNavBar features={dummyData.features} dispatch={this.props.dispatch}/>
+          <h3>Expenses Tracker</h3>
           <div>
-            <h3>Current Expenses</h3>
-            {this.state.expenses.map((item) => {
-              return <ExpenseEntry expense={item} key={item.id} payer={this.findUser(item.userId)}/>
-            })}
-            <h3>Total Cost</h3> ${this.state.totalExpense}
+            <ExpenseInput usersOnTrip={this.state.usersOnTrip} fetchExpenses={this.fetchExpenses.bind(this)} />
+            <hr />
+            <div>
+              <h3>Current Expenses</h3>
+              {this.state.expenses.map((item) => {
+                return <ExpenseEntry expense={item} key={item.id} payer={this.findUser(item.userId)}/>
+              })}
+              <h3>Total Cost</h3> 
+              <div class="ExpenseEntry">${this.state.totalExpense}</div>
+            </div>
           </div>
-        </div>
-        <div>
-          <button className="btnback" onClick={this.handleBack.bind(this)}>Back</button>
-        </div>
-      </div>
-		)
+        </Col>
+      </Row>
+    )
 	}
 }
